allow report to be specified from command line
