+++ lib/mergometer/reports/merge_trend_report_entry.rb
@@ -1,0 +1,20 @@
+module Mergometer
+  module Reports
+    class MergeTrendReportEntry
+      attr_reader :date
+
+      def initialize(date, prs)
+        @date = date
+        @prs = prs
+      end
+
+      def median_merge_time
+        Math.median prs.map(&:merge_time)
+      end
+
+      private
+
+        attr_reader :prs
+    end
+  end
+end
