+++ lib/mergometer/reports/review_report.rb
@@ -2,47 +2,49 @@ require "mergometer/report"
 require "mergometer/reports/review_report_entry"
 
 module Mergometer
-  class ReviewReport < Report
-    def render
-      super
-      puts "PRs awaiting review (#{prs_awaiting_review.size}): " + prs_awaiting_review.join(", ")
-    end
-
-    private
-
-
-      def prs_awaiting_review
-        prs.find_all(&:unreviewed?).map do |pr|
-          "##{pr.number}"
-        end
+  module Reports
+    class ReviewReport < Report
+      def render
+        super
+        puts "PRs awaiting review (#{prs_awaiting_review.size}): " + prs_awaiting_review.join(", ")
       end
 
-      def fields
-        [:user, :reviews]
-      end
+      private
 
-      def filter
-        "repo:#{repo} type:pr updated:>=#{1.week.ago.to_date}"
-      end
 
-      def fields_to_preload
-        %i(reviewers pr_data)
-      end
-
-      def entries
-        review_data.map do |user, reviews|
-          ReviewReportEntry.new(user: user, reviews: reviews)
-        end
-      end
-
-      def review_data
-        prs.inject({}) do |result, pr|
-          pr.reviewers.each do |reviewer|
-            result[reviewer] ||= 0
-            result[reviewer] += 1
+        def prs_awaiting_review
+          prs.find_all(&:unreviewed?).map do |pr|
+            "##{pr.number}"
           end
-          result
         end
-      end
+
+        def fields
+          [:user, :reviews]
+        end
+
+        def filter
+          "repo:#{repo} type:pr updated:>=#{1.week.ago.to_date}"
+        end
+
+        def fields_to_preload
+          %i(reviewers pr_data)
+        end
+
+        def entries
+          review_data.map do |user, reviews|
+            ReviewReportEntry.new(user: user, reviews: reviews)
+          end
+        end
+
+        def review_data
+          prs.inject({}) do |result, pr|
+            pr.reviewers.each do |reviewer|
+              result[reviewer] ||= 0
+              result[reviewer] += 1
+            end
+            result
+          end
+        end
+    end
   end
 end
