+++ lib/mergometer/reports/review_report_entry.rb
@@ -1,10 +1,12 @@
 module Mergometer
-  class ReviewReportEntry
-    attr_accessor :user, :reviews
+  module Reports
+    class ReviewReportEntry
+      attr_accessor :user, :reviews
 
-    def initialize(user:, reviews:)
-      @user = user
-      @reviews = reviews
+      def initialize(user:, reviews:)
+        @user = user
+        @reviews = reviews
+      end
     end
   end
 end
