+++ lib/mergometer/reports/weekly_report_entry.rb
@@ -1,19 +1,21 @@
 module Mergometer
-  class WeeklyReportEntry
+  module Reports
+    class WeeklyReportEntry
 
-    attr_accessor :week
+      attr_accessor :week
 
-    def initialize(week:, prs:)
-      @week = week
-      @prs = prs
+      def initialize(week:, prs:)
+        @week = week
+        @prs = prs
+      end
+
+      def count
+        prs.size
+      end
+
+      private
+
+        attr_accessor :prs
     end
-
-    def count
-      prs.size
-    end
-
-    private
-
-      attr_accessor :prs
   end
 end
